requirements.txt
Django==3.0.4
django-bulma==0.6.0
django-jquery==3.1.0
django-jquery-ui==1.11.4.1
django-silk==4.0.1
requests==2.23.0
SPARQLWrapper==1.8.2

networkx~=2.4